feat: redactor play sound, add sounds to kit
