+++ src/assets/scss/_variables.scss
@@ -14,7 +14,7 @@ $color-sea-buckthorn: #fbb034;
 $color-school-bus-yellow: #fd0;
 $color-emperor: #525050;
 $color-red: #f00;
-$color-apple: #70a36c;
+$color-apple: #8db600;
 $color-ocean-green: #3aa76d;
 $color-chestnut: #b54a4a;
 $color-cinnamon: #e25822;
