+++ src/common/helpers.js
@@ -1,5 +1,8 @@
 import { ExternalUrls, LocalStorageKeys, DefaultValues } from './constants';
 import * as _ from 'lodash';
+import correctSound from '../assets/audio/correct.wav';
+import wrongSound from '../assets/audio/wrong.wav';
+import charSound from '../assets/audio/char.wav';
 
 export function buildUrl(...args) {
 	return args.join('');
@@ -39,4 +42,24 @@ export function getStreak(array) {
 	});
 
 	return _.max(result);
+}
+
+export function playSound(soundSrc) {
+	const sound = new Audio(soundSrc);
+	sound.play();
+}
+
+export function playWrong() {
+	const sound = new Audio(wrongSound);
+	sound.play();
+}
+
+export function playCorrect() {
+	const sound = new Audio(correctSound);
+	sound.play();
+}
+
+export function playChar() {
+	const sound = new Audio(charSound);
+	sound.play();
 }
