--- src/common/utils.js
@@ -1,4 +1,0 @@
-export function playSound(soundSrc) {
-	const sound = new Audio(soundSrc);
-	sound.play();
-}
