+++ src/components/Games/Kit/Kit.module.scss
@@ -55,6 +55,7 @@
 }
 
 .shuffledChar {
+	position: relative;
 	display: flex;
 	width: 40px;
 	height: 40px;
@@ -73,7 +74,16 @@
 	}
 
 	&:focus {
-		background-color: red;
+		outline: none;
+
+		&::after {
+			content: '';
+			position: absolute;
+			width: 45px;
+			height: 45px;
+			border: 1px solid $color-apple;
+			border-radius: 8px;
+		}
 	}
 
 	&[aria-selected='true'] {
